Tarea07-PracticaEnEquipo/Ejercicio_03/requirements.txt
requests==2.32.5
Pillow==11.3.0
numpy==2.2.5